Rename BaseLogger to WriterLogger to help the creation of other providers
